pyproject.toml
[tool.poetry]
name = "my-wallet"
version = "0.1.0"
description = ""
authors = ["NickolayBabulich <nickolas.grn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
